Removed hiding/showing of diff view
Traded hiding/showing of diff view in for
a fixed size of the diff view. Previously the
size of the diff view got automatically adapted
to the width/height of its contained text. This
made quick browsing of changes hard, because the
position of the diff view changed too.

Unfortunatly enabling hiding/showing AND fixed
size of the diff view doesn't work properly
with the current version of cursive.